+++ src/datacite/__init__.py
@@ -1,0 +1,7 @@
+"""Create and manage DOIs using the DataCite API
+
+This package provides a command line script to mint and manage
+DataCite DOIs using the DataCite REST API.
+"""
+
+from ._meta import version as __version__